Make host networking configurable via OpenSearchCluster
OpenSearch operator doesn't offer the possibility to configure host networking.
So at first, we hardcoded the value to true to make OpenSearch work on our pipelines.

With the arrival of Cilium, we want to have the possibility to configure freely the host networking of
an OpenSearch cluster.

We can override the hostNetwork attribute in 4 different places:
  * bootstrap
  * dashboards
  * general
  * by nodePool

If the hostNetwork attribute is set in the nodePool/dashboards and in the general config,
the nodePool/dashboards configuration takes priority.

The configuration still works like before if the hostNetwork attribute is not
set at all (hostNetwork is set to true by default).

Signed-off-by: f.bouzghaia <f.bouzghaia@criteo.com>